backend/scraper/requirements.txt
scrapy==2.11.0
beautifulsoup4==4.12.2
requests==2.31.0
kafka-python==2.0.2
celery==5.3.4
redis==5.0.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0